Rename calculator class
